Move favorites context provider to entities
