+++ superhero-directory/src/app/providers/providers.tsx
@@ -1,4 +1,4 @@
-import { FavoritesProvider } from '~features/favorites';
+import { FavoritesProvider } from '~entities/favorites';
 
 import { QueryClientProvider } from './react-query';
