+++ superhero-directory/src/entities/favorites/index.ts
